refactor(use_case): tighten file boundaries, drop dead code, dedupe helpers
Restructure the use_case package to eliminate dead code, dedupe repeated
patterns, and clarify module responsibilities. Net -1250 lines across
16 files; 739 tests pass.

- Delete introspector.py (646 lines) — never called from the production
  path; server_graphql + compose_schema already provided the same info.
- Delete types.py — merge UseCaseAppConfig into manager.py (sole consumer).
- Delete apply_selection + 3 helpers from selection.py — only tests used
  them; compose.py uses build_subset_model directly.
- Extract iter_use_case_methods into business.py — collapses the
  getattr + kind-filter pattern duplicated in compose_schema and
  server_graphql.
- Extract is_from_context_annotation into context.py — consolidates 4
  inline copies.
- Move SDL helpers (method_sdl et al.) from server_graphql.py to
  compose_schema.py — they operate on the schema compose_schema builds.
- Move compose_introspect + is_introspection_query into new
  introspection.py — execution pipeline in compose.py is data-only.
- Cache build_compose_schema result on UseCaseResources.compose_schema
  so introspection + describe_compose_method don't rebuild it per call.
- Rename _method_sdl → method_sdl (cross-module use made the underscore
  misleading).
- Drop dead build_compose_schema import from compose.py.
