+++ pydantic_resolve/use_case/__init__.py
@@ -38,9 +38,8 @@ Example:
 
 from pydantic_resolve.use_case.business import UseCaseService
 from pydantic_resolve.use_case.context import FromContext
-from pydantic_resolve.use_case.manager import UseCaseResources
+from pydantic_resolve.use_case.manager import UseCaseAppConfig, UseCaseResources
 from pydantic_resolve.use_case.server_graphql import create_use_case_graphql_mcp_server
-from pydantic_resolve.use_case.types import UseCaseAppConfig
 
 __all__ = [
     "create_use_case_graphql_mcp_server",
