+++ pydantic_resolve/use_case/compose_schema.py
@@ -26,14 +26,11 @@ queries to ``compose_introspect``.
 from __future__ import annotations
 
 import inspect
-from typing import Annotated, Any, Callable, get_args, get_origin
+from typing import Any, Callable
 
 from graphql import (
     GraphQLArgument,
-    GraphQLBoolean,
     GraphQLField,
-    GraphQLFloat,
-    GraphQLInt,
     GraphQLNonNull,
     GraphQLObjectType,
     GraphQLSchema,
@@ -41,8 +38,8 @@ from graphql import (
 )
 
 from pydantic_resolve.graphql.type_converter import pydantic_to_graphql_type
-from pydantic_resolve.use_case.business import USE_CASE_METHODS_ATTR
-from pydantic_resolve.use_case.context import FromContext
+from pydantic_resolve.use_case.business import iter_use_case_methods
+from pydantic_resolve.use_case.context import is_from_context_annotation
 from pydantic_resolve.utils.types import _resolve_function_type_hints, get_return_annotation
 
 
@@ -106,16 +103,10 @@ def _build_service_object_type(
     Each ``@query`` / ``@mutation`` method on the service becomes a
     GraphQLField; ``cls`` and ``FromContext`` params are skipped.
     """
-    methods: dict[str, dict[str, Any]] = getattr(
-        service_cls, USE_CASE_METHODS_ATTR, {}
-    )
-
     fields: dict[str, GraphQLField] = {}
-    for method_name, meta in methods.items():
-        kind = meta.get("kind", "query") if isinstance(meta, dict) else "query"
-        if kind == "mutation" and not enable_mutation:
-            continue
-
+    for method_name, _kind, meta in iter_use_case_methods(
+        service_cls, enable_mutation=enable_mutation
+    ):
         method = meta["method"]
         func = getattr(method, "__func__", method)
         return_anno = get_return_annotation(method)
@@ -164,7 +155,7 @@ def _build_method_args(
         if name == "cls":
             continue
         anno = hints.get(name, param.annotation)
-        if _is_from_context_param(anno):
+        if is_from_context_annotation(anno):
             continue
         if anno is inspect.Parameter.empty or anno is None:
             anno = str  # unknown → default to String
@@ -185,19 +176,109 @@ def _build_method_args(
     return args
 
 
-def _is_from_context_param(annotation: Any) -> bool:
-    if annotation is None or annotation is inspect.Parameter.empty:
-        return False
-    if get_origin(annotation) is not Annotated:
-        return False
-    return any(isinstance(arg, FromContext) for arg in get_args(annotation))
-
-
 def _docstring_or_meta(meta: dict[str, Any]) -> str | None:
     desc = meta.get("description") if isinstance(meta, dict) else None
     if desc:
         return desc
     return None
+
+
+# ============================================================================
+# SDL rendering (focused per-method view, used by describe_compose_method)
+# ============================================================================
+
+
+def method_sdl(schema: Any, service_name: str, method_name: str) -> str | None:
+    """Focused SDL for one method: signature + return type's transitive closure.
+
+    Returns a GraphQL SDL string showing:
+    - The method signature (args + return type) as a comment header
+    - Full type definitions for the return DTO and every nested DTO
+      reachable through its fields (handles cycles)
+
+    Returns ``None`` if the method or its return type can't be located
+    in ``schema`` (e.g. scalar returns have no nested types to print).
+
+    Callers should pass the cached ``app.compose_schema`` rather than
+    rebuilding on each call.
+    """
+    from graphql.utilities import print_type
+
+    service_field = schema.query_type.fields.get(service_name)
+    if service_field is None:
+        return None
+    service_type = _unwrap_type(service_field.type)
+    method_field = service_type.fields.get(method_name) if service_type else None
+    if method_field is None:
+        return None
+
+    # Collect reachable object types from the return type
+    reachable: dict[str, Any] = {}
+    _collect_reachable_types(method_field.type, schema.type_map, reachable)
+
+    sdl_parts: list[str] = []
+    # Method signature as a comment header
+    args_sdl = ", ".join(
+        f"{name}: {_graphql_type_to_sdl(arg.type)}"
+        for name, arg in method_field.args.items()
+    )
+    sdl_parts.append(
+        f"# {service_name}.{method_name}({args_sdl}): "
+        f"{_graphql_type_to_sdl(method_field.type)}"
+    )
+    for type_name, type_def in sorted(reachable.items()):
+        sdl_parts.append(print_type(type_def))
+    return "\n\n".join(sdl_parts)
+
+
+def _unwrap_type(type_ref: Any) -> Any:
+    """Peel NonNull / List wrappers to get the underlying named type."""
+    while hasattr(type_ref, "of_type"):
+        type_ref = type_ref.of_type
+    return type_ref
+
+
+def _graphql_type_to_sdl(type_ref: Any) -> str:
+    """Render a graphql-core type reference as SDL syntax.
+
+    ``GraphQLNonNull(GraphQLList(GraphQLNonNull(GraphQLObjectType(X))))``
+    → ``[X!]!``
+    """
+    type_name = getattr(type_ref, "name", None)
+    if type_name is not None:
+        return type_name
+    inner = getattr(type_ref, "of_type", None)
+    if inner is None:
+        return str(type_ref)
+    inner_sdl = _graphql_type_to_sdl(inner)
+    # GraphQLNonNull wraps as ``T!``; GraphQLList wraps as ``[T]``
+    class_name = type(type_ref).__name__
+    if class_name == "GraphQLNonNull":
+        return f"{inner_sdl}!"
+    if class_name == "GraphQLList":
+        return f"[{inner_sdl}]"
+    return inner_sdl
+
+
+def _collect_reachable_types(
+    type_ref: Any, type_map: dict[str, Any], seen: dict[str, Any]
+) -> None:
+    """DFS-walk a graphql-core type reference, recording every object type.
+
+    Used to build the transitive closure of types reachable from a
+    method's return type — so the SDL response shows nested DTOs
+    without dumping the entire schema.
+    """
+    from graphql.type import GraphQLObjectType
+
+    core = _unwrap_type(type_ref)
+    name = getattr(core, "name", None)
+    if name is None or name in seen:
+        return
+    if isinstance(core, GraphQLObjectType):
+        seen[name] = core
+        for field in core.fields.values():
+            _collect_reachable_types(field.type, type_map, seen)
 
 
 __all__ = ["build_compose_schema"]
