+++ pydantic_resolve/use_case/manager.py
@@ -7,13 +7,39 @@ subclasses. Follows the same pattern as GraphQL MCP's MultiAppManager.
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Awaitable, Callable
+from typing import Any, Awaitable, Callable
 
-from pydantic_resolve.use_case.introspector import ServiceIntrospector
-from pydantic_resolve.use_case.types import UseCaseAppConfig
+from graphql import GraphQLSchema
+from pydantic import BaseModel
 
-if TYPE_CHECKING:
-    from pydantic_resolve.use_case.business import UseCaseService
+from pydantic_resolve.use_case.business import UseCaseService
+from pydantic_resolve.use_case.compose_schema import build_compose_schema
+
+
+class UseCaseAppConfig(BaseModel):
+    """Configuration for a UseCase application in MCP server.
+
+    Attributes:
+        name: Application name (required)
+        services: List of UseCaseService subclasses for this app (required)
+        description: Optional application description
+        enable_mutation: Whether mutation methods are exposed via MCP (default: True).
+            When False, mutation methods are hidden from
+            ``describe_compose_schema``, ``describe_compose_method``,
+            and ``compose_query``.
+        context_extractor: Optional callback that extracts request-scoped context
+            (e.g. user identity from Authorization header). Receives the FastMCP
+            Context object and returns a dict injected as ``_context`` parameter
+            into UseCaseService methods. Can be sync or async.
+    """
+
+    model_config = {"arbitrary_types_allowed": True}
+
+    name: str
+    services: list[type["UseCaseService"]]
+    description: str | None = None
+    enable_mutation: bool = True
+    context_extractor: Callable[[Any], dict | Awaitable[dict]] | None = None
 
 
 @dataclass
@@ -23,15 +49,15 @@ class UseCaseResources:
     Attributes:
         name: Application name
         description: Application description
-        introspector: ServiceIntrospector for method discovery
         services: Mapping of service name to service class
+        compose_schema: Cached graphql-core schema (built once at registration).
         context_extractor: Optional callback to extract request-scoped context
     """
 
     name: str
     description: str
-    introspector: ServiceIntrospector
     services: dict[str, type["UseCaseService"]] = field(default_factory=dict)
+    compose_schema: GraphQLSchema | None = None
     context_extractor: Callable[[Any], dict | Awaitable[dict]] | None = field(default=None)
     enable_mutation: bool = True
 
@@ -75,7 +101,7 @@ class UseCaseManager:
     This manager handles:
     - Registration of multiple apps from UseCaseAppConfig list
     - App lookup by name (with case-insensitive fallback)
-    - Resource creation for each app (ServiceIntrospector, service mapping)
+    - Resource creation for each app (service mapping)
 
     Each app is independent and contains its own group of services.
     """
@@ -103,7 +129,7 @@ class UseCaseManager:
             config: Application configuration
 
         Returns:
-            UseCaseResources instance with introspector and service mapping
+            UseCaseResources instance with service mapping
         """
         name = config.name
         description = config.description or ""
@@ -114,17 +140,17 @@ class UseCaseManager:
         for svc in services:
             service_map[svc.__name__] = svc
 
-        # Create introspector for this app's services
-        introspector = ServiceIntrospector(services)
-
-        return UseCaseResources(
+        resources = UseCaseResources(
             name=name,
             description=description,
-            introspector=introspector,
             services=service_map,
             context_extractor=config.context_extractor,
             enable_mutation=config.enable_mutation,
         )
+        # Eager-build the graphql-core schema once; it never changes after
+        # registration. All introspection / SDL paths read from this cache.
+        resources.compose_schema = build_compose_schema(resources)
+        return resources
 
     def _register_app(self, resources: UseCaseResources) -> None:
         """Register an app's resources.
