+++ pydantic_resolve/use_case/selection.py
@@ -2,9 +2,9 @@
 
 This module implements a lightweight, DTO-only projection layer for the
 field-selection level of ``compose_query``.  It intentionally reuses the
-GraphQL ``QueryParser``/``FieldSelection`` structures while avoiding
-ERD-specific response building behavior such as relationships,
-pagination, and FK fields.
+GraphQL ``FieldSelection`` structure while avoiding ERD-specific
+response building behavior such as relationships, pagination, and FK
+fields.
 """
 
 from __future__ import annotations
@@ -14,65 +14,19 @@ import typing
 from types import UnionType as _UnionType
 from typing import Any, get_args, get_origin
 
-from pydantic import BaseModel, ConfigDict, Field, TypeAdapter, create_model
+from pydantic import BaseModel, ConfigDict, Field, create_model
 from pydantic_core import PydanticUndefined
 
-from pydantic_resolve.graphql.exceptions import QueryParseError
-from pydantic_resolve.graphql.query_parser import QueryParser
 from pydantic_resolve.graphql.types import FieldSelection
 from pydantic_resolve.utils.class_util import safe_issubclass
 from pydantic_resolve.utils.types import get_core_types
 
 
 _UNION_ORIGINS = (typing.Union, _UnionType)
-_RESULT_FIELD = "__result"
 
 
 class SelectionError(ValueError):
     """Raised when a UseCase MCP selection is invalid."""
-
-
-def apply_selection(result: Any, return_annotation: Any, selection: str) -> Any:
-    """Project a UseCase result into a dynamic Pydantic subset model.
-
-    The returned value remains a Pydantic model/list of Pydantic models/None;
-    final serialization is still handled by the existing MCP response layer.
-    """
-    field_selection = parse_selection(selection)
-    root_model, root_annotation = _extract_root_model(return_annotation, result)
-    subset_model = build_subset_model(root_model, field_selection)
-
-    if result is None:
-        return None
-
-    try:
-        selected_annotation = _replace_model_type(root_annotation, subset_model)
-        return TypeAdapter(selected_annotation).validate_python(result)
-    except Exception as e:  # pragma: no cover - exact pydantic errors vary
-        raise SelectionError(f"Failed to apply selection: {e}") from e
-
-
-def parse_selection(selection: str) -> FieldSelection:
-    """Parse a rootless GraphQL-like selection into a FieldSelection tree."""
-    if not selection or not selection.strip():
-        raise SelectionError("selection cannot be empty")
-
-    query = f"{{ {_RESULT_FIELD} {selection} }}"
-    try:
-        parsed = QueryParser().parse(query)
-    except QueryParseError as e:
-        raise SelectionError(str(e)) from e
-
-    root = parsed.field_tree.get(_RESULT_FIELD)
-    if root is None:
-        raise SelectionError("selection could not be parsed")
-    if root.arguments:
-        raise SelectionError("selection arguments are not supported")
-    if not root.sub_fields:
-        raise SelectionError("selection must include at least one field")
-
-    _reject_arguments(root)
-    return root
 
 
 def build_subset_model(
@@ -119,68 +73,6 @@ def build_subset_model(
         __config__=ConfigDict(from_attributes=True, arbitrary_types_allowed=True),
         **field_definitions,
     )
-
-
-def _reject_arguments(selection: FieldSelection, path: str = "") -> None:
-    if selection.arguments:
-        location = path or _RESULT_FIELD
-        raise SelectionError(f"selection arguments are not supported at '{location}'")
-    if not selection.sub_fields:
-        return
-    for field_name, sub_selection in selection.sub_fields.items():
-        child_path = f"{path}.{field_name}" if path else field_name
-        _reject_arguments(sub_selection, child_path)
-
-
-def _extract_root_model(
-    return_annotation: Any,
-    result: Any,
-) -> tuple[type[BaseModel], Any]:
-    core_type = _get_pydantic_core_type(return_annotation)
-    if core_type is not None:
-        return core_type, return_annotation
-
-    runtime_annotation = _infer_runtime_annotation(result)
-    runtime_core_type = _get_pydantic_core_type(runtime_annotation)
-    if runtime_core_type is not None:
-        return runtime_core_type, runtime_annotation
-
-    raise SelectionError(
-        "selection is only supported for Pydantic return types "
-        "(BaseModel, list[BaseModel], or optional variants)"
-    )
-
-
-def _infer_runtime_annotation(result: Any) -> Any:
-    if isinstance(result, BaseModel):
-        return result.__class__
-
-    if not isinstance(result, list):
-        return None
-
-    item_type = None
-    saw_none = False
-    for item in result:
-        if item is None:
-            saw_none = True
-            continue
-
-        if not isinstance(item, BaseModel):
-            return None
-
-        current_type = item.__class__
-        if item_type is None:
-            item_type = current_type
-        elif current_type is not item_type:
-            return None
-
-    if item_type is None:
-        return None
-
-    if saw_none:
-        return list[item_type | None]
-
-    return list[item_type]
 
 
 def _get_pydantic_core_type(annotation: Any) -> type[BaseModel] | None:
@@ -244,20 +136,6 @@ def _strip_annotated(annotation: Any) -> Any:
             break
         annotation = args[0]
     return annotation
-
-
-def _is_optional_annotation(annotation: Any) -> bool:
-    origin = get_origin(annotation)
-    return origin in _UNION_ORIGINS and any(arg is type(None) for arg in get_args(annotation))
-
-
-def _strip_optional(annotation: Any) -> Any:
-    if not _is_optional_annotation(annotation):
-        return annotation
-    non_none = [arg for arg in get_args(annotation) if arg is not type(None)]
-    if len(non_none) != 1:
-        return annotation
-    return non_none[0]
 
 
 def _is_list_annotation(annotation: Any) -> bool:
