+++ tests/use_case/test_compose_mcp.py
@@ -19,8 +19,8 @@ from pydantic import BaseModel
 from pydantic_resolve import query, mutation
 from pydantic_resolve.use_case.business import UseCaseService
 from pydantic_resolve.use_case.context import FromContext
+from pydantic_resolve.use_case.manager import UseCaseAppConfig
 from pydantic_resolve.use_case.server_graphql import create_use_case_graphql_mcp_server
-from pydantic_resolve.use_case.types import UseCaseAppConfig
 
 
 # ──────────────────────────────────────────────────
